pyproject.toml
[project]
name = "mrzvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy==2.4.0",
    "pillow~=12.0.0",
    "tabulate>=0.9.0",
]

exclude = [
    ".venv",
    "dist",
    "build",
]
